refactor(db): centralize OIDC tables into core schema
WS-A (oidc slice) of the module-data-centralization plan. All 10 OIDC tables
(jwks, device_codes, oauth_clients, oauth_refresh/access_tokens, oauth_consents,
cli_refresh_tokens, oidc_end_user_profiles, application_smtp_configs,
application_social_providers) move from the OIDC module into the core schema;
the system pipeline creates them at boot. Module keeps its behavior (routes,
Better Auth plugins, RBAC, realm resolver).

- packages/db/src/schema/oidc.ts: table defs (relative core imports). The 4
  raw-SQL CHECK constraints (oauth_clients level + signup_role,
  application_smtp_configs secure_mode, application_social_providers provider)
  reproduced via Drizzle check() so regen keeps them.
- 0003_fold_oidc_tables.sql: idempotent adoption (to_regclass guard) carrying
  the oauth_clients level-immutability TRIGGER + function verbatim (not
  expressible in Drizzle), then drops the stale __drizzle_migrations_oidc.
  Parity gate: pg_dump diff vs the module migrations = constraint-NAME diffs
  only (Drizzle naming), zero structural diff. Trigger + idempotency + CHECK
  enforcement verified on real Postgres.
- Better Auth tables auto-register via the core schema barrel (auth.ts spreads
  `import * as schema`); oidc keeps drizzleSchemas() for now (redundant,
  harmless) — its removal + the contract change land in WS-D.
- oidc module: ~20 files rewired to import tables from @appstrate/db/schema;
  init() drops applyMigrations. Module schema.ts + drizzle/ deleted.
- oidc suite green (430 pass); webhooks green (61); db + api typecheck green.
