+++ apps/api/src/modules/oidc/test/integration/services/device-flow-realm.test.ts
@@ -29,7 +29,7 @@ import { flushRedis } from "../../../../../../test/helpers/redis.ts";
 import oidcModule from "../../../index.ts";
 import { resetOidcGuardsLimiters } from "../../../auth/guards.ts";
 import { ensureCliClient } from "../../../services/ensure-cli-client.ts";
-import { deviceCode } from "../../../schema.ts";
+import { deviceCode } from "@appstrate/db/schema";
 
 const app = getTestApp({ modules: [oidcModule] });
